refactor: generalise TypeZ machinery
We now have a general "nested zipper" definition, which will be used to
focus on kinds-inside-types{,-inside-exprs} later.
