+++ primer/src/Primer/Zipper/Type.hs
@@ -1,22 +1,10 @@
-{-# LANGUAGE FunctionalDependencies #-}
-
 -- | This module contains the zipper type @TypeZ@, and functions for
 --  operating on them.
 module Primer.Zipper.Type (
   TypeZip,
   TypeZip',
-  IsZipper (asZipper),
-  focus,
-  target,
-  _target,
-  replace,
   focusOnKind,
   focusOnTy,
-  top,
-  up,
-  down,
-  left,
-  right,
   farthest,
   search,
   FoldAbove,
@@ -40,17 +28,8 @@ import Data.Data (Data)
 import Data.Generics.Uniplate.Data ()
 import Data.Generics.Uniplate.Zipper (
   Zipper,
-  replaceHole,
-  zipper,
  )
-import Data.Generics.Uniplate.Zipper qualified as Z
 import Data.Set qualified as S
-import Optics (
-  over,
-  view,
- )
-import Optics.Lens (Lens', equality', lens)
-import Optics.Traversal (traverseOf)
 import Primer.Core.Meta (
   HasID (..),
   ID,
@@ -64,6 +43,15 @@ import Primer.Core.Type (
   TypeMeta,
  )
 import Primer.Name (Name)
+import Primer.Zipper.Nested (
+  IsZipper,
+  down,
+  focus,
+  left,
+  right,
+  target,
+  up,
+ )
 
 type KindZip' c = Zipper (Kind' c) (Kind' c)
 
@@ -71,47 +59,6 @@ type TypeZip' b c = Zipper (Type' b c) (Type' b c)
 
 -- | An ordinary zipper for 'Type's
 type TypeZip = TypeZip' TypeMeta ()
-
--- | We want to use up, down, left, right, etc. on 'ExprZ' and 'TypeZ',
--- despite them being very different types. This class enables that, by proxying
--- each method through to the underlying Zipper.
--- @za@ is the user-facing type, i.e. 'ExprZ' or 'TypeZ'.
--- @a@ is the type of targets of the internal zipper, i.e. 'Expr' or 'Type'.
-class Data a => IsZipper za a | za -> a where
-  asZipper :: Lens' za (Z.Zipper a a)
-
-instance Data a => IsZipper (Z.Zipper a a) a where
-  asZipper = equality'
-
-target :: IsZipper za a => za -> a
-target = Z.hole . view asZipper
-
--- | A 'Lens' for the target of a zipper
-_target :: IsZipper za a => Lens' za a
-_target = lens target (flip replace)
-
-up :: IsZipper za a => za -> Maybe za
-up = traverseOf asZipper Z.up
-
-down :: (IsZipper za a) => za -> Maybe za
-down = traverseOf asZipper Z.down
-
-left :: IsZipper za a => za -> Maybe za
-left = traverseOf asZipper Z.left
-
-right :: IsZipper za a => za -> Maybe za
-right = traverseOf asZipper Z.right
-
-top :: IsZipper za a => za -> za
-top = farthest up
-
--- | Convert a normal 'Expr' or 'Type' to a cursored one, focusing on the root
-focus :: (Data a) => a -> Zipper a a
-focus = zipper
-
--- | Replace the node at the cursor with the given value.
-replace :: (IsZipper za a) => a -> za -> za
-replace = over asZipper . replaceHole
 
 -- | Focus on the node with the given 'ID', if it exists in the kind
 focusOnKind ::
